Add note about blank values in colors.json
I made this mistake a few times in the beginning. If you leave out a key-value
pair in colors.json, Go's text/template engine will leave behind the string
"<no value>" in its place. Obviously, those angle brackets will ruin the XML,
and you'll get a nasty error when the theme is loaded.

Basically, every key that is referenced in the template theme must be included
in colors.json.