Refactor definitions in colors.json
Since pallor has only five highlight colors, the colors tend to be repeated a
lot. When you want to change a color, you have to update all the uses in
colors.json.

With this commit, colors are now defined in a single JSON object, and then
referenced throughout the various highlighted scopes. Now you can easily
change an individual color without changing the scopes it applies to, or vice
versa.