Add highlighting for markup.math scope
I do a lot of work with pandoc inline math, which is wrapped in $ signs, so
this commit allows pallor to highlight those things in orange.

No grammar implements this scope currently. I have an experimental extension
for Markdown Extended which provides it, but I'm not really satisfied with the
behavior. I think, instead of just highlighting it all orange, it should be
highlighted as TeX (which makes sense, since the contents of $ are TeX math).

For now, I'll leave this commit as an interim solution. Once I have a grammar
that implements this more effectively, we'll see where I want to take this
change.